pyproject.toml
[project]
name = "playground-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
    "pynput>=1.7.7",
]
